Scenes complete
Interriror, exterrior, and survival mode levels all complete. Everything
(should be) is a static object that needs to be.

there are some colliders to mark the boundries of the map, which MAY
interfere with enemy AI spawning in survival mode, but Ian knows how to
make special nav-meshes for ai